requirements.txt
Django==3.2.5
django-admin==2.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
drf-extra-fields==3.4.0
drf-yasg==1.20.0
pandas==1.5.3
psycopg2==2.9.5
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.0/en_core_web_md-3.7.0.tar.gz#egg=en_core_web_md==3.7.0
#https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0.tar.gz#egg=fr_core_news_md==3.7.0
https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.7.0/fr_core_news_lg-3.7.0-py3-none-any.whl#sha256=c2d98eb9073ff0f3d865179448e7f1868f27b2643c0e7215acc2438cd9eae554
SQLAlchemy==2.0.23
langdetect==1.0.9
django-filter
openpyxl
